Encouragement Text Bot/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.7.1
aiosignal==1.2.0
anyio==3.4.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==3.0.1
attrs==21.4.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cssselect==1.1.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
discord==1.7.3
discord.py==1.7.3
EbookLib==0.17.1
elastic-transport==8.1.1
elasticsearch==8.1.2
entrypoints==0.3
fake-useragent==0.1.11
fileupload==0.1.5
Flask==2.1.2
fonttools==4.28.5
frozenlist==1.3.0
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
graphviz==0.19.1
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
json5==0.9.6
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.5
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
librosa==0.9.1
llvmlite==0.38.0
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.7
noisereduce==2.0.0
numba==0.55.1
numpy==1.21.2
oauthlib==3.2.0
packaging==21.3
pandas==1.3.3
pandoc==2.0.1
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.0.0
pip-autoremove==0.10.0
plumbum==1.7.2
ply==3.11
pooch==1.6.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
pyasn1==0.4.8
pycares==4.1.2
pycparser==2.21
pydub==0.25.1
Pygments==2.11.1
pyparsing==3.0.6
pyquery==1.4.3
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-socks==2.0.3
pytz==2021.3
pywin32==302
pywinpty==1.1.6
pyzmq==22.3.0
regex==2021.11.10
replit==3.2.4
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
schedule==1.1.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.3.1
statsmodels==0.13.2
terminado==0.12.1
testpath==0.5.0
textblob==0.17.1
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tweepy==4.8.0
twint==2.1.20
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.1.2
widgetsnbextension==3.5.2
wordcloud==1.8.1
yarl==1.7.2
zipp==3.8.0